requirements.txt
nbtlib>=1.11.2
PyYAML>=6.0